@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateSubjects < ActiveRecord::Migration
+class CreateSubjects < ActiveRecord::Migration[4.2]
   def change
     create_table :subjects do |t|
       t.string :name, null: false
