@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Migration to create roles table and relevant fields
-class CreateRoles < ActiveRecord::Migration
+class CreateRoles < ActiveRecord::Migration[4.2]
   def change
     create_table :roles do |t|
       t.string :name, null: false
