@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Migration to create api_subject_roles table and relevant fields
-class CreateAPISubjectRoles < ActiveRecord::Migration
+class CreateAPISubjectRoles < ActiveRecord::Migration[4.2]
   def change
     create_table :api_subject_roles do |t|
       t.integer :role_id, null: false
