@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Migration to create subject_roles table and relevant fields
-class CreateSubjectRoles < ActiveRecord::Migration
+class CreateSubjectRoles < ActiveRecord::Migration[4.2]
   def change
     create_table :subject_roles do |t|
       t.integer :role_id, null: false
