@@ -34,13 +34,18 @@ module Authentication
       end
     end
 
-    def finish(_env)
+    def finish(env)
+      url = env['rack.session']['return_url'].to_s
+      env['rack.session'].delete('return_url')
+
+      return redirect_to(url) if url.present?
+
       redirect_to(latest_snapshots_path)
     end
 
     # :nocov:
     def logout(env)
-      env['rack.session'].clear
+      env['rack.session'].destroy
 
       return redirect_to('/') unless Rails.env.production?
 
